refactor(idp/authn): update components on new dependencies
