+++ identity-provider/authentication/src/main/java/io/wliamp/idp/authn/compo/handler/ResponseHandler.java
@@ -7,7 +7,7 @@ import reactor.core.publisher.Mono;
 
 @Component
 public class ResponseHandler {
-    public Mono<ServerResponse> buildTokenResponse(Tokens token) {
+    public Mono<ServerResponse> buildTokensResponse(Tokens token) {
         return ServerResponse.ok()
                 .header("X-Access-Token", token.access())
                 .header("X-Refresh-Token", token.refresh())
