+++ identity-provider/authentication/src/main/java/io/wliamp/idp/authn/service/db/AccService.java
@@ -1,4 +1,4 @@
-package io.wliamp.idp.authn.service.data;
+package io.wliamp.idp.authn.service.db;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
